day 11 part 2

+++ src/test/java/adventOfCode2020/day11/Puzzle11Test.java
@@ -42,7 +42,7 @@ public class Puzzle11Test {
     private void assertSolve2(boolean isExampleData) {
         AbstractPuzzle puzzle = isExampleData ? puzzleExampleData : puzzleRealData;
         Object answerObj = puzzle.solve2();
-        long answer = (Long) answerObj;
+        int answer = (Integer) answerObj;
         long expected = puzzle.getExpectedAnswer2();
         assertTrue("Expected " + expected + " but got " + answer, answer == expected);
     }
